feat(enquete): add individual information forms (#1622)
* feat(mandataire-agrements): add mandataire agrements table

* feat(mandataire-agrements): add mandataire_id to mandataire agrements

* feat(mandataire-agrements): add mandataire agrement comps

* feat(mandataire-agrements): update mandataire infos page with agrement

* feat(ind-agrement): update model

* feat(exercice): add migration / components

* feat(formation): add migration / components

* feat(agrement): handle update

* feat(validation):  update

* feat(mandataire-exercice): handle form

* feat(mandataire-formation): handle form

* feat(mandataire-info): add text

* feat(indi): create objects in signup

* chore(test): update snapshot

* chore(test): update test

* chore(dep): remove @socialgouv/emjpm-ui-core

* chore(test): update snapshot

* feat(enquete): style form

* feat(formation): add view component

* feat(agrement): add view component

* feat(exercice): add view component

* refactor(enquete): move FieldLabel in its own file

* feat(secretariat): change column name

* chore(test): update snapshot

+++ packages/api/__test__/e2e/inscription/POST_inscription.test.js
@@ -46,9 +46,24 @@ afterEach(async () => {
     // Not all users have a `user_tis` relation.
   }
 
-  await knex("mandataires")
+  const mandataire = await knex("mandataires")
     .where({ user_id: user.id })
-    .delete();
+    .first();
+
+  if (mandataire) {
+    await knex("individuel_agrements")
+      .where({ mandataire_id: mandataire.id })
+      .delete();
+    await knex("individuel_formations")
+      .where({ mandataire_id: mandataire.id })
+      .delete();
+    await knex("individuel_exercices")
+      .where({ mandataire_id: mandataire.id })
+      .delete();
+    await knex("mandataires")
+      .where({ id: mandataire.id })
+      .delete();
+  }
 
   await knex("magistrat")
     .where({ user_id: user.id })
